requirements.txt
asgiref==3.2.3
Django==3.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
python-sqlite-orm==0.0.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0